Moved the DAO classes to respective sub-packages
