+++ src/main/java/in/gagan/base/framework/dao/CountryDAO.java
@@ -21,6 +21,7 @@ package in.gagan.base.framework.dao;
 
 import java.util.Optional;
 
+import in.gagan.base.framework.dao.base.BaseDAO;
 import in.gagan.base.framework.dto.location.ZipcodeDTO;
 import in.gagan.base.framework.entity.location.City;
 import in.gagan.base.framework.entity.location.Country;
