+++ src/main/java/in/gagan/base/framework/dao/user/VerificationTokenDAO.java
@@ -17,10 +17,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package in.gagan.base.framework.dao;
+package in.gagan.base.framework.dao.user;
 
 import java.util.Optional;
 
+import in.gagan.base.framework.dao.base.BaseDAO;
 import in.gagan.base.framework.entity.VerificationToken;
 
 /**
