+++ src/main/java/in/gagan/base/framework/dao/user/VerificationTokenDAOImpl.java
@@ -17,26 +17,27 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package in.gagan.base.framework.dao;
+package in.gagan.base.framework.dao.user;
 
 import java.util.List;
 import java.util.Optional;
 
 import javax.persistence.Query;
 
+import in.gagan.base.framework.dao.base.AbstractBaseDAO;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.CollectionUtils;
 
-import in.gagan.base.framework.entity.ForgotPasswordToken;
+import in.gagan.base.framework.entity.VerificationToken;
 
 /**
- * This class provides CRUD operations on the FORGOT_PASSWORD_TOKEN table using DAO pattern.
+ * This class provides CRUD operations on the VERIFICATION_TOKEN table using DAO pattern.
  * 
  * @author gaganthind
  *
  */
 @Repository
-public class ForgotPasswordTokenDAOImpl extends AbstractBaseDAO<ForgotPasswordToken, Long> implements ForgotPasswordTokenDAO {
+public class VerificationTokenDAOImpl extends AbstractBaseDAO<VerificationToken, Long> implements VerificationTokenDAO {
 	
 	private final static String WHERE_TOKEN_CLAUSE = " where token = :token ";
 	private final static String TOKEN = "token";
@@ -45,17 +46,17 @@ public class ForgotPasswordTokenDAOImpl extends AbstractBaseDAO<ForgotPasswordTo
 	 * Method used to fetch a token record.
 	 * 
 	 * @param token - token to fetch record
-	 * @return Optional<ForgotPasswordToken> - ForgotPasswordToken record
+	 * @return Optional<VerificationToken> - VerificationToken record
 	 */
 	@SuppressWarnings("unchecked")
 	@Override
-	public Optional<ForgotPasswordToken> fetchByToken(String token) {
+	public Optional<VerificationToken> fetchByToken(String token) {
 		String selectQuery = new StringBuilder(LITERAL_FROM).append(getTableName()).append(WHERE_TOKEN_CLAUSE).toString();
 		Query query = entityManager.createQuery(selectQuery);
 		query.setParameter(TOKEN, token);
-		List<ForgotPasswordToken> forgotPasswordTokens = (List<ForgotPasswordToken>) query.getResultList();
+		List<VerificationToken> verificationTokens = (List<VerificationToken>) query.getResultList();
 		
-		return !CollectionUtils.isEmpty(forgotPasswordTokens) ? Optional.of(forgotPasswordTokens.get(0)) : Optional.empty();
+		return !CollectionUtils.isEmpty(verificationTokens) ? Optional.of(verificationTokens.get(0)) : Optional.empty();
 	}
-
+	
 }
