+++ src/main/java/in/gagan/base/framework/service/admin/AdminServiceImpl.java
@@ -20,7 +20,7 @@
 package in.gagan.base.framework.service.admin;
 
 import in.gagan.base.framework.component.AdminAccount;
-import in.gagan.base.framework.dao.UserDAO;
+import in.gagan.base.framework.dao.user.UserDAO;
 import in.gagan.base.framework.entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
