+++ src/main/java/in/gagan/base/framework/service/user/LoginAttemptServiceImpl.java
@@ -27,7 +27,7 @@ import org.springframework.stereotype.Service;
 
 import in.gagan.base.framework.component.PasswordProps;
 import in.gagan.base.framework.constant.ApplicationConstants;
-import in.gagan.base.framework.dao.UserDAO;
+import in.gagan.base.framework.dao.user.UserDAO;
 import in.gagan.base.framework.entity.User;
 import in.gagan.base.framework.util.ExceptionHelperUtil;
